+++ backend/orchestrator/esbuild.config.mjs
@@ -1,0 +1,9 @@
+import * as esbuild from 'esbuild';
+
+await esbuild.build({
+  entryPoints: ['src/index.ts'],
+  bundle: true,
+  platform: 'node',
+  outfile: 'dist/index.js',
+  external: ['@fastify/swagger', '@fastify/swagger-ui'],
+});
