+++ backend/orchestrator/package.json
@@ -7,7 +7,7 @@
     "test": "echo \"Error: no test specified\" && exit 1",
     "lint": "tsc",
     "dev": "run-p dev:*",
-    "build": "esbuild src/index.ts --bundle --platform=node --outfile=dist/index.js",
+    "build": "node ./esbuild.config.mjs",
     "start": "node --env-file=.env dist/index.js",
     "execute": "pnpm build && pnpm start",
     "dev:tsc": "tsc --watch --preserveWatchOutput",
@@ -18,22 +18,21 @@
   "author": "",
   "license": "ISC",
   "devDependencies": {
+    "@fastify/swagger": "^8.14.0",
+    "@fastify/swagger-ui": "^4.0.0",
     "@types/node": "^20.14.2",
     "esbuild": "^0.21.5",
+    "fastify": "^4.28.0",
+    "fastify-type-provider-zod": "^1.2.0",
     "npm-run-all": "^4.1.5",
-    "typescript": "^5.4.5"
+    "typescript": "^5.4.5",
+    "zod": "^3.22.4",
+    "zod-to-json-schema": "^3.23.1"
   },
   "dependencies": {
     "@langchain/community": "^0.2.11",
     "@langchain/core": "^0.2.6",
     "@langchain/langgraph": "^0.0.23",
-    "@langchain/openai": "^0.1.3",
-    "zod": "^3.22.4",
-    "zod-to-json-schema": "^3.23.1"
-  },
-  "pnpm": {
-    "overrides": {
-      "@langchain/core": "0.2.6"
-    }
+    "@langchain/openai": "^0.1.3"
   }
 }
