+++ backend/orchestrator/tsconfig.json
@@ -22,5 +22,5 @@
   },
   // Include the necessary files for your project
   "include": ["**/*.ts"],
-  "exclude": ["node_modules"]
+  "exclude": ["node_modules", "esbuild.config.mjs"]
 }
