+++ package.json
@@ -38,5 +38,10 @@
     "globals": "^15.4.0",
     "prettier": "^3.3.1",
     "typescript-eslint": "^7.12.0"
+  },
+  "pnpm": {
+    "overrides": {
+      "@langchain/core": "0.2.6"
+    }
   }
 }
